ml/requirements.txt
ultralytics>=8.2.0
onnx>=1.15.0
numpy>=1.24.0
Pillow>=10.0.0
pyyaml>=6.0.1
